move callbacks to separate file; add icons, style
